Aqua refactor: use const in aqua.constants.py  (#849)
