+++ ads/aqua/common/utils.py
@@ -27,7 +27,7 @@ from ads.aqua.common.errors import (
     AquaRuntimeError,
     AquaValueError,
 )
-from ads.aqua.constants import SERVICE_MANAGED_CONTAINER_URI_SCHEME
+from ads.aqua.constants import *
 from ads.aqua.data import AquaResourceIdentifier
 from ads.common.auth import AuthState, default_signer
 from ads.common.extended_enum import ExtendedEnumMeta
@@ -38,50 +38,6 @@ from ads.config import AQUA_SERVICE_MODELS_BUCKET, CONDA_BUCKET_NS, TENANCY_OCID
 from ads.model import DataScienceModel, ModelVersionSet
 
 logger = logging.getLogger("ads.aqua")
-
-UNKNOWN = ""
-UNKNOWN_DICT = {}
-README = "README.md"
-LICENSE_TXT = "config/LICENSE.txt"
-DEPLOYMENT_CONFIG = "deployment_config.json"
-COMPARTMENT_MAPPING_KEY = "service-model-compartment"
-CONTAINER_INDEX = "container_index.json"
-EVALUATION_REPORT_JSON = "report.json"
-EVALUATION_REPORT_MD = "report.md"
-EVALUATION_REPORT = "report.html"
-UNKNOWN_JSON_STR = "{}"
-CONSOLE_LINK_RESOURCE_TYPE_MAPPING = dict(
-    datasciencemodel="models",
-    datasciencemodeldeployment="model-deployments",
-    datasciencemodeldeploymentdev="model-deployments",
-    datasciencemodeldeploymentint="model-deployments",
-    datasciencemodeldeploymentpre="model-deployments",
-    datasciencejob="jobs",
-    datasciencejobrun="job-runs",
-    datasciencejobrundev="job-runs",
-    datasciencejobrunint="job-runs",
-    datasciencejobrunpre="job-runs",
-    datasciencemodelversionset="model-version-sets",
-    datasciencemodelversionsetpre="model-version-sets",
-    datasciencemodelversionsetint="model-version-sets",
-    datasciencemodelversionsetdev="model-version-sets",
-)
-FINE_TUNING_RUNTIME_CONTAINER = "iad.ocir.io/ociodscdev/aqua_ft_cuda121:0.3.17.20"
-DEFAULT_FT_BLOCK_STORAGE_SIZE = 750
-DEFAULT_FT_REPLICA = 1
-DEFAULT_FT_BATCH_SIZE = 1
-DEFAULT_FT_VALIDATION_SET_SIZE = 0.1
-
-HF_MODELS = "/home/datascience/conda/pytorch21_p39_gpu_v1/"
-MAXIMUM_ALLOWED_DATASET_IN_BYTE = 52428800  # 1024 x 1024 x 50 = 50MB
-JOB_INFRASTRUCTURE_TYPE_DEFAULT_NETWORKING = "ME_STANDALONE"
-NB_SESSION_IDENTIFIER = "NB_SESSION_OCID"
-LIFECYCLE_DETAILS_MISSING_JOBRUN = "The asscociated JobRun resource has been deleted."
-READY_TO_DEPLOY_STATUS = "ACTIVE"
-READY_TO_FINE_TUNE_STATUS = "TRUE"
-AQUA_GA_LIST = ["id19sfcrra6z"]
-AQUA_MODEL_TYPE_SERVICE = "service"
-AQUA_MODEL_TYPE_CUSTOM = "custom"
 
 
 class LifecycleStatus(str, metaclass=ExtendedEnumMeta):
@@ -143,8 +99,6 @@ LIFECYCLE_DETAILS_MAPPING = {
     JobRun.LIFECYCLE_STATE_FAILED: "The evaluation failed.",
     JobRun.LIFECYCLE_STATE_NEEDS_ATTENTION: "Missing jobrun information.",
 }
-SUPPORTED_FILE_FORMATS = ["jsonl"]
-MODEL_BY_REFERENCE_OSS_PATH_KEY = "artifact_location"
 
 
 def random_color_generator(word: str):
